requirements.txt
mysql-connector-python==8.4.0
matplotlib==3.9.0
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
